+++ app/src/main/java/com/example/runtracker/gallery/GalleryImageDetailsActivity.kt
@@ -1,4 +1,4 @@
-package com.example.runtracker
+package com.example.runtracker.gallery
 
 import android.content.Context
 import android.content.ContextWrapper
@@ -8,6 +8,7 @@ import android.os.Bundle
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
+import com.example.runtracker.R
 import com.squareup.picasso.Picasso
 import java.io.File
