+++ app/src/main/java/com/example/runtracker/gallery/Images.kt
@@ -1,3 +1,3 @@
-package com.example.runtracker
+package com.example.runtracker.gallery
 
 class Images(var path: String, var data: String)
